Code code refactoring.
Preparing for distributed builds support.
Use Jenkins FilePath to manange local nodes paths.

+++ src/main/java/caphyon/jenkins/advinst/AdvinstCommands.java
@@ -5,10 +5,9 @@
  */
 package caphyon.jenkins.advinst;
 
-import java.io.File;
-import java.io.FileOutputStream;
+import hudson.FilePath;
+
 import java.io.IOException;
-import java.io.OutputStreamWriter;
 import java.util.List;
 
 /**
@@ -23,43 +22,36 @@ public class AdvinstCommands
    * Executes a batch of commands on the specified AIP file.
    *
    * @param aAdvinstPath Path to advanced installer command line tool
-   * @param aAipFile path to the AIP project file on which the commands are
-   * ran
-   * @param aCommands List of commands
-   * @param aLog Output parameter containing the output log of the Advanced
-   * Installer commands line tool
+   * @param aAipFile     path to the AIP project file on which the commands are
+   *                     ran
+   * @param aCommands    List of commands
+   * @param aLog         Output parameter containing the output log of the Advanced
+   *                     Installer commands line tool
    * @return true if all commands were successfully executed
    * @throws AdvinstException
    */
-  public static boolean executeCommandsBatch(final String aAdvinstPath, final String aAipFile, final List<String> aCommands, StringBuilder aLog) throws AdvinstException
+  public static boolean executeCommandsBatch(final FilePath aAdvinstPath, final FilePath aAipFile, final FilePath aWorkFolder,
+                                             final List<String> aCommands, StringBuilder aLog) throws AdvinstException
   {
-    File commandsFile = null;
+    FilePath commandsFile = null;
     try
     {
-      commandsFile = createAicFile(aCommands);
+      commandsFile = aWorkFolder;
+      commandsFile = aWorkFolder.createTempFile("aic", "");
+      createAicFile(commandsFile, aCommands);
+      final String executeCommand = String.format("\"%s\" /execute \"%s\" \"%s\"",
+        aAdvinstPath,
+        aAipFile,
+        commandsFile);
+      Command cmd = new Command(executeCommand);
+      int errCode = cmd.execute();
+      aLog.append(cmd.getOutput());
+      commandsFile.delete();
+      return 0 == errCode;
     }
     catch (IOException ex)
     {
       throw new AdvinstException("Failed to create commands file. Exception: " + ex.getMessage(), ex);
-    }
-
-    final String executeCommand = String.format("\"%s\" /execute \"%s\" \"%s\"",
-            aAdvinstPath,
-            aAipFile,
-            commandsFile.getAbsolutePath());
-
-    Command cmd = new Command(executeCommand);
-    int errCode;
-    try
-    {
-      errCode = cmd.execute();
-      aLog.append(cmd.getOutput());
-      commandsFile.delete();
-      return errCode == 0;
-    }
-    catch (IOException ex)
-    {
-      throw new AdvinstException("Failed to execute commands batch. Exception: " + ex.getMessage(), ex);
     }
     catch (InterruptedException ex)
     {
@@ -67,40 +59,35 @@ public class AdvinstCommands
     }
     finally
     {
-      commandsFile.delete();
-    }
-  }
-
-  private static File createAicFile(final List<String> aCommands) throws IOException
-  {
-
-    File aicFile = File.createTempFile("aic", null);
-    FileOutputStream outStream = new FileOutputStream(aicFile);
-    OutputStreamWriter writer = null;
-    try
-    {
-      writer = new OutputStreamWriter(outStream, "UTF-16");
-      writer.write(AdvinstConsts.AdvinstAicHeader + "\r\n");
-      for (String command : aCommands)
+      if (null != commandsFile)
       {
-        writer.write(command + "\r\n");
-      }
-    }
-    finally
-    {
-      try
-      {
-
-        if (null != writer)
+        try
         {
-          writer.close();
+          commandsFile.delete();
+        }
+        catch (IOException ex)
+        {
+          throw new AdvinstException(ex.getMessage(), ex);
+        }
+        catch (InterruptedException ex)
+        {
+          throw new AdvinstException(ex.getMessage(), ex);
         }
       }
-      catch (IOException ex)
-      {
-      }
+    }
+  }
+
+  private static boolean createAicFile(final FilePath cmdsFilePath, final List<String> aCommands) throws IOException, InterruptedException
+  {
+    String fileContent = AdvinstConsts.AdvinstAicHeader + "\r\n";
+    for (String command : aCommands)
+    {
+      fileContent += command;
+      fileContent += "\r\n";
     }
 
-    return aicFile;
+    cmdsFilePath.write(fileContent, "UTF-16");
+    return true;
   }
+
 }
